221121 두번째 브랜치_main controller 분리_ 검토부탁드립니다.
